requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
lxml==4.7.1
mysqlclient==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-docx==0.8.11
PyYAML==6.0
requests==2.26.0
SQLAlchemy==1.4.29
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
zipp==3.6.0
